Notes_Keeper-main/backend/requirements.txt
flask
flask-cors
flask-pymongo
flask-jwt-extended
flask-bcrypt
python-dotenv
dnspython
bson